code formatting fixed

+++ asp_app/Services/Operations/DoubleMultiplication.cs
@@ -1,12 +1,12 @@
-﻿
-namespace ComputingService.Services.Operations
+﻿namespace ComputingService.Services.Operations
 {
-    public class DoubleMultiplication : BinaryOperationBase<double>
-    {
-	    public override string Name => "mul";
-	    public override double Compute()
-	    {
-		    return (double)Left * Right;
-	    }
-    }
+	public class DoubleMultiplication : BinaryOperationBase<double>
+	{
+		public override string Name => "mul";
+
+		public override double Compute()
+		{
+			return Left * Right;
+		}
+	}
 }
